requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# LangChain Core & Integrations
langchain
langchain-community
langchain-openai
langchain-chroma

# Vector Database & Data Handling
chromadb
pysqlite3-binary  # Important for some cloud hosting environments
pandas
numpy

# Utilities
python-dotenv
pydantic